Refactor duration handling from `Days` to `Day`
- Updated the usage of `Days` to `Day` across multiple modules including `period.rs`, `scales.rs`, `time.rs`, and FFI bindings.
- Adjusted related tests to ensure consistency with the new `Day` type.
- Modified function signatures and internal logic to accommodate the change in duration representation.
- Ensured that all assertions in tests reflect the new `Day` type for duration comparisons.

+++ tempoch-ffi/src/period.rs
@@ -182,7 +182,7 @@ mod tests {
             end_mjd: 51_546.5,
         };
         let qty = tempoch_period_mjd_duration_qty(p);
-        assert_eq!(qty.unit, UnitId::Day);
+        assert_eq!(qty.unit, UnitId::Day as u32);
         assert!((qty.value - 2.0).abs() < 1e-12);
     }
